% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{callTessi}
\alias{callTessi}
\title{Send a Request to the Tessitura API}
\usage{
callTessi(host, basePath, method, credentials, request_type = "GET",
  data, flatten = TRUE)
}
\arguments{
\item{host}{character}

\item{basePath}{character}

\item{method}{character}

\item{credentials}{character}

\item{request_type}{character}

\item{data}{list}

\item{flatten}{logical}
}
\value{
list
}
\description{
Send a Request to the Tessitura API
}
\examples{
#' callTessi('mytessi.tessituranetwork.com/', 'TessituraService', '/Diagnostics/Status', 'mybase64credentials')
}
