% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{createTessituraService}
\alias{createTessituraService}
\title{Create a new Tessitura Service object}
\usage{
createTessituraService(tessituraUrl, credentials)
}
\arguments{
\item{tessituraUrl}{The base url for your Tessitura REST API}

\item{credentials}{A base64 encded character string. Can be created with createCredentials()}
}
\value{
a TessituraService S4 object
}
\description{
Create a new Tessitura Service object
}
