% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessituraR-package.R
\docType{package}
\name{tessituraR}
\alias{tessituraR}
\alias{tessituraR-package}
\title{tessituraR: Interacting With the Tessitura API}
\description{
A package to help you work with the Tessitura API. At the moment,
this is just a baby package to help with API calls, but the goal is to
create a package that will help to do all kinds of data analysis with
knowledge of Tessitura data structures.
}
