% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{callTessi}
\alias{callTessi}
\title{Send a Request to the Tessitura API}
\usage{
callTessi(host, basePath, resource, credentials, request_type = "GET",
  data, flatten = TRUE)
}
\arguments{
\item{host}{The host name for your Tessitura API}

\item{basePath}{The base path for your Tessitura API. eg., TessituraService}

\item{resource}{The resource to be request from the API. eg., Diagnostics/Status}

\item{credentials}{A base64 encoded character string. Can be created with createCredentials()}

\item{request_type}{An http verb. Currently on GET and POST are supported}

\item{data}{The data object for POST.}

\item{flatten}{If true, a data frame will be returned of the results. If false, the result object will be returned.}
}
\value{
A data frame or list of the result
}
\description{
Send a Request to the Tessitura API
}
\examples{
#' callTessi('mytessi.tessituranetwork.com/', 'TessituraService', '/Diagnostics/Status', 'mybase64credentials')
}
