% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{createCredentials}
\alias{createCredentials}
\title{Create Tessitura API Credentials}
\usage{
createCredentials(username, usergroup, location, password)
}
\arguments{
\item{username}{character}

\item{usergroup}{character}

\item{location}{character}

\item{password}{character}
}
\value{
character
}
\description{
Create Tessitura API Credentials
}
\examples{
#' createCredentials(username="creif", usergroup="admin", location="MET95", password="impresario")
}
