% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{createCredentials}
\alias{createCredentials}
\title{Create Tessitura API Credentials}
\usage{
createCredentials(username, usergroup, location, password)
}
\arguments{
\item{username}{Your Tessitura user name}

\item{usergroup}{Your Tessitura user group}

\item{location}{Your Tessitura machine location}

\item{password}{Your Tessitura password}
}
\value{
A base64 encoded character string for authorization
}
\description{
Create Tessitura API Credentials
}
\examples{
#' createCredentials(username="creif", usergroup="admin", location="MET95", password="impresario")
}
